requirements.txt
transformers
optimum
openai-whisper
